requirements.txt
fastapi
uvicorn[standard]
python-multipart
PyMuPDF
python-docx
spacy